+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/InvoiceEditPanel.java
@@ -118,8 +118,8 @@ public final class InvoiceEditPanel extends EntityEditPanel {
 		public Selector apply(EntitySearchModel searchModel) {
 			TableSelector selector = EntitySearchField.tableSelector(searchModel);
 			selector.table().columnModel().visible().set(Customer.LASTNAME, Customer.FIRSTNAME, Customer.EMAIL);
-			selector.table().sortModel().setSortOrder(Customer.LASTNAME, ASCENDING);
-			selector.table().sortModel().addSortOrder(Customer.FIRSTNAME, ASCENDING);
+			selector.table().model().sorter().setSortOrder(Customer.LASTNAME, ASCENDING);
+			selector.table().model().sorter().addSortOrder(Customer.FIRSTNAME, ASCENDING);
 			selector.preferredSize(new Dimension(500, 300));
 
 			return selector;