+++ demos/manual/src/main/java/is/codion/framework/demos/manual/notes/NotesDemo.java
@@ -192,7 +192,7 @@ public final class NotesDemo {
 							// Note.CREATED is excluded by default since it is not updatable.
 							.editable(attributes -> attributes.remove(Note.UPDATED)));
 			// Configure the table and columns
-			table().sortModel().setSortOrder(Note.CREATED, SortOrder.DESCENDING);
+			table().model().sorter().setSortOrder(Note.CREATED, SortOrder.DESCENDING);
 			table().setAutoResizeMode(JTable.AUTO_RESIZE_SUBSEQUENT_COLUMNS);
 			FilterTableColumnModel<Attribute<?>> columnModel = table().columnModel();
 			columnModel.column(Note.NOTE).setPreferredWidth(280);