+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityTablePanelTest.java
@@ -166,7 +166,7 @@ public class EntityTablePanelTest {
 		assertEquals(Employee.NAME, orderBy.orderByColumns().get(1).column());
 
 		EntityTablePanel tablePanel = new EntityTablePanel(tableModel);
-		tablePanel.table().sortModel().setSortOrder(Employee.NAME, SortOrder.ASCENDING);
+		tablePanel.table().model().sorter().setSortOrder(Employee.NAME, SortOrder.ASCENDING);
 
 		orderBy = tableModel.queryModel().orderBy().get();
 		//still default order by for entity
@@ -182,8 +182,8 @@ public class EntityTablePanelTest {
 		assertTrue(orderBy.orderByColumns().get(0).ascending());
 		assertEquals(Employee.NAME, orderBy.orderByColumns().get(0).column());
 
-		tablePanel.table().sortModel().setSortOrder(Employee.HIREDATE, SortOrder.DESCENDING);
-		tablePanel.table().sortModel().addSortOrder(Employee.NAME, SortOrder.ASCENDING);
+		tablePanel.table().model().sorter().setSortOrder(Employee.HIREDATE, SortOrder.DESCENDING);
+		tablePanel.table().model().sorter().addSortOrder(Employee.NAME, SortOrder.ASCENDING);
 
 		orderBy = tableModel.queryModel().orderBy().get();
 		assertEquals(2, orderBy.orderByColumns().size());
@@ -192,7 +192,7 @@ public class EntityTablePanelTest {
 		assertTrue(orderBy.orderByColumns().get(1).ascending());
 		assertEquals(Employee.NAME, orderBy.orderByColumns().get(1).column());
 
-		tablePanel.table().sortModel().clear();
+		tablePanel.table().model().sorter().clear();
 		orderBy = tableModel.queryModel().orderBy().get();
 		//back to default order by for entity
 		assertEquals(2, orderBy.orderByColumns().size());